requirements.txt
presidio-analyzer[transformers]
presidio-analyzer[stanza]
presidio-anonymizer
streamlit>=1.37.0
streamlit-tags
pandas>=1.0.4
python-dotenv
st-annotated-text
torch>=2.8.0
flair>=0.15.0
openai
azure-ai-textanalytics
pymongo[srv]
